Create DateChecker.js
A date checking function that compares two dates. Accepts two dates as strings in YYYY-MM-DD format. The dates are converted to arrays and then compared to each other based on indices to determine which date comes earliest. If the date entered as the first parameter is later than the date entered as the second parameter, or if they are the same date, false is returned. True is returned if the first date is earlier than the second date.