Create palindromeChecker.js
A palindrome checker that accepts a single string parameter. Converts the string to an array and reverses it with a for loop. The original array and the reversed array are then compared to determine if the string is a palindrome. If the two arrays are the same, true is returned, otherwise false is returned.