Create GeneTranslator.js
A series of functions for translating DNA directly to protein, DNA to RNA and RNA to Protein. The translate function accepts a dna sequence string containing 'a', 't', 'g', and 'c'. The string is then converted to an array and then iterated by three's to determine a codon. The codon codes for a specific amino acid. A switch statement determines which codon codes for which amino acid and adds it to a protein sequence string. The function convertNucleotides accepts a string of either DNA or RNA. For a DNA string, all T's will be converted to U's. For RNA, all U's will be converted to T's. The function RNAtoProtein accepts an RNA string and then converts it to DNA and then inserts the resulting string into the translate function to code for protein.